solve error in cheque

+++ account_cheque/models/models.py
@@ -24,16 +24,20 @@ class account_cheque(models.Model):
     @api.depends('amount','currency_id')
     def compute_amount_arabic_word(self):
         for rec in self:
+
             rec.total_amount_arabic_word = ''
             if rec.currency_id:
                 amount = round(rec.amount, 4)
                 text = rec.currency_id.amount_to_text(amount)
                 integer = int(amount)
-                decimal = round((amount - integer), 2) * 100
+                decimal = int(round((amount - integer), 2) * 100)
+
                 riyal = num2words(integer, lang='ar')
                 halala = num2words(decimal, lang='ar')
                 text_ar = riyal
+
                 text_ar += " " + str(rec.currency_id.currency_unit_arabic)
+
                 if decimal:
                     text_ar += ' و'
                     text_ar += halala
@@ -41,6 +45,7 @@ class account_cheque(models.Model):
 
                     # total_amount=str(text_ar).split()
                 amount_text = text_ar
+
                 line_list = str(text_ar).split()
                 if 'عشرين' in line_list:
                     amount_text = str(str(text_ar).replace("عشرين", "عشرون"))
